Fixed uiValidateWatch
For some reason I can't checkout this branch locally.

Everything seems to work now, just need unit tests and peer review.

This has a breaking API change, you now must wrap expressions in strings.
This is to allow 2-step parsing. Once for the object of validators, and again to validate.